criando rotas de consulta todos , consultar por id , editar , criar e excluir do cardapio

+++ src/main/java/com/rasmoo/api/rasfood/repository/CardapioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardapioRepository extends JpaRepository<Cardapio,Long> {
+public interface CardapioRepository extends JpaRepository<Cardapio,Integer> {
 }
